# Conflicts:
#	scripts/o3de/o3de/enable_gem.py
#	scripts/o3de/o3de/gem_properties.py
#	scripts/o3de/o3de/project_properties.py
#	scripts/o3de/o3de/utils.py
#	scripts/o3de/tests/test_gem_properties.py
#	scripts/o3de/tests/test_project_manager_interface.py

+++ Code/Framework/AzNetworking/AzNetworking/TcpTransport/TcpSocket.cpp
@@ -109,7 +109,7 @@ namespace AzNetworking
 
     int32_t TcpSocket::SendInternal(const uint8_t* data, uint32_t size) const
     {
-        const int32_t sentBytes = send(aznumeric_cast<int32_t>(m_socketFd), (const char*)data, size, 0);
+        const int32_t sentBytes = static_cast<int32_t>(send(aznumeric_cast<int32_t>(m_socketFd), (const char*)data, size, 0));
 
         if (sentBytes < 0)
         {
@@ -126,7 +126,7 @@ namespace AzNetworking
 
     int32_t TcpSocket::ReceiveInternal(uint8_t* outData, uint32_t size) const
     {
-        const int32_t receivedBytes = recv(aznumeric_cast<int32_t>(m_socketFd), (char*)outData, (int32_t)size, 0);
+        const int32_t receivedBytes = static_cast<int32_t>(recv(aznumeric_cast<int32_t>(m_socketFd), (char*)outData, (int32_t)size, 0));
 
         if (receivedBytes < 0)
         {
